increase size of QF fields

+++ IPTables.Net/Conntrack/ConntrackQueryFilter.cs
@@ -9,8 +9,8 @@ namespace IPTables.Net.Conntrack
     [StructLayout(LayoutKind.Sequential, Pack=4, Size=16)]
     public struct ConntrackQueryFilter
     {
-        public UInt16 Key;
-        public UInt16 Max;
+        public Int32 Key;
+        public Int32 Max;
         public Int32 CompareLength;
         public IntPtr Compare;
     }
